requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anaconda-navigator==2.4.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
binaryornot==0.4.4
bkcharts==0.2
branca==0.6.0
bs4==0.0.1
cachetools==5.3.0
census==0.8.19
certifi==2022.12.7
citipy==0.0.5
clikit==0.6.2
clyent==1.2.2
cmdstanpy==1.1.0
conda==23.1.0
conda-build==3.23.3
conda-repo-cli==1.0.27
config==0.5.1
constantly==15.1.0
convertdate==2.4.0
crashtest==0.3.1
daal4py==2023.0.2
datashape==0.5.4
dnspython==2.2.1
ephem==4.1.4
Flask-PyMongo==2.3.0
flatbuffers==23.1.21
folium==0.14.0
frozenlist==1.3.3
gast==0.4.0
glob2==0.7
gmaps==0.9.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
HeapDict==1.0.1
hijri-converter==2.2.4
holidays==0.21
html5lib==1.1
httpstan==4.9.1
intervaltree==3.0.2
ipython-genutils==0.2.0
jdcal==1.4.1
kdtree==0.16
keras==2.11.0
Keras-Preprocessing==1.1.2
kneed==0.8.2
korean-lunar-calendar==0.3.1
libclang==15.0.6.1
llvmlite==0.39.1
LunarCalendar==0.0.9
marshmallow==3.19.0
mccabe==0.6.1
mkl-service==2.4.0
mrjob==0.7.4
multidict==6.0.4
multipledispatch==0.6.0
munkres==1.1.4
navigator-updater==0.3.0
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.1
opt-einsum==3.3.0
outcome==1.2.0
pandas==1.5.3
pastel==0.2.1
pep8==1.7.1
prophet==1.1.2
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyct==0.4.6
pycurl==7.45.1
PyDispatcher==2.0.5
pylev==1.4.0
PyMeeus==0.5.12
pymongo==4.3.3
pysimdjson==5.0.2
pystan==2.19.1.1
pytest==7.2.1
python-dotenv==0.21.0
python-lsp-server==1.2.4
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn-intelex==20230131.200356
scipy==1.10.1
selenium==4.7.2
service-identity==18.1.0
setuptools-git==1.2
sip==4.19.13
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
splinter==0.18.1
SQLAlchemy==2.0.7
TBB==0.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-macos==2.11.0
tensorflow-metal==0.7.1
termcolor==2.2.0
text-unidecode==1.3
trio==0.22.0
trio-websocket==0.9.2
webargs==8.2.0
webdriver-manager==3.8.5
webencodings==0.5.1
wsproto==1.2.0
yarl==1.8.2
zstandard==0.19.0